fix custom hook location to query pokemons
